Added serialize functions for knownGene
git-svn-id: https://compbio.charite.de/public-svn/NGSanalysis@749 a7cbf6ab-7213-4657-a3ac-d80d95007299

+++ Exomizer/src/exomizer/exome/Variant.java
@@ -119,7 +119,7 @@ public class Variant implements Comparable<Variant>, Constants {
     /**
      * @return an byte representation of the chromosome, e.g., 1,2,...,22 (note: X=23, Y=24, MT=25).
      */
-    public byte getChromosomeAsByte() { return (byte)chromosome; }
+    public byte getChromosomeAsByte() { return chromosome; }
    
 
     public boolean is_X_chromosomal() { return this.chromosome == X_CHROMOSOME;  }
